+++ Src/Frames/StageSelectFrame.h
@@ -89,7 +89,7 @@ public:
 			frame->type = _type;
 		}
 
-		inline void SetNowStageSelect(unsigned long _nowSelectStage)
+		inline void SetNowStageSelect(size_t _nowSelectStage)
 		{
 			frame->nowSelectStage = _nowSelectStage;
 		}
@@ -118,7 +118,7 @@ public:
 
 	protected:
 
-		inline unsigned long GetNowSelectCount()
+		inline size_t GetNowSelectCount()
 		{
 			return frame->nowSelectStage;
 		}
@@ -217,7 +217,7 @@ private:
 
 	ChD3D11::Shader::BaseDrawSprite11 spriteShader;
 
-	unsigned long nowSelectStage = 0;
+	size_t nowSelectStage = 0;
 
 	ChD3D11::Texture11 notImageTexture;
